Separate module for examples. This change helps spot problems when some types are available only on runtime classpath, and not compile classpath
