+++ example/build.gradle
@@ -1,0 +1,8 @@
+apply plugin: 'java'
+
+dependencies {
+    implementation project(":reactor-java")
+    implementation project(":codec-jackson")
+    implementation 'io.rsocket:rsocket-transport-netty'
+}
+
