+++ settings.gradle
@@ -6,4 +6,5 @@ include 'codec-jackson'
 include 'codec-jackson-binary'
 include 'codec-proto'
 include 'contract'
+include 'example'
